requirements.txt

torch==2.2.2
torchvision==0.17.2
ultralytics>=8.2.0
streamlit>=1.30.0

opencv-python==4.9.0.80
pillow>=10.0.0

numpy==1.26.4
pyyaml>=6.0.0
scipy>=1.11.0